Add cdb service to get configuration data from the webserver with the refresh rate
The cdb service can request configuration data of the integrated alarm system from the webserver.
Currently the cdb service is initialized in the available alarm service
and is able to retreive the configuration refresh rate. This refresh rate value is used to
set a local validation for the alarms in terms of delay.